requirements.txt
django
django-heroku
pillow
gunicorn
djangorestframework-jwt
djangorestframework
django-cors-headers
django-rest-auth
django-allauth
python-dotenv
django-storages[dropbox]
